+++ src/wf_psf/data/centroids.py
@@ -8,9 +8,78 @@ A module with utils to handle PSF centroids.
 
 import numpy as np
 import scipy.signal as scisig
-from wf_psf.data.data_preprocessing import shift_x_y_to_zk1_2_wavediff
+from wf_psf.data.data_handler import extract_star_data
+from fractions import Fraction
+from wf_psf.data.data_zernike_utils import shift_x_y_to_zk1_2_wavediff
+import tensorflow as tf
 from typing import Optional
 
+
+def compute_centroid_correction(model_params, data, batch_size: int=1) -> np.ndarray:
+    """Compute centroid corrections using Zernike polynomials.
+
+    This function calculates the Zernike contributions required to match the centroid
+    of the WaveDiff PSF model to the observed star centroids, processing in batches.
+
+    Parameters
+    ----------
+    model_params : RecursiveNamespace
+        An object containing parameters for the PSF model, including pixel sampling and initial centroid window parameters.
+
+    data : DataConfigHandler
+        An object containing training and test datasets, including observed PSFs
+        and optional star masks.
+
+    batch_size : int, optional
+        The batch size to use when processing the stars. Default is 16.
+
+
+    Returns
+    -------
+    zernike_centroid_array : np.ndarray
+         A 2D NumPy array of shape `(n_stars, 3)`, where `n_stars` is the number of 
+        observed stars. The array contains the computed Zernike contributions, 
+        with zero padding applied to the first column to ensure a consistent shape.
+    """
+    star_postage_stamps = extract_star_data(data=data, train_key="noisy_stars", test_key="stars")
+    
+    # Get star mask catalogue only if "masks" exist in both training and test datasets
+    star_masks = (
+    extract_star_data(data=data, train_key="masks", test_key="masks")
+    if (
+        data.training_data.dataset.get("masks") is not None 
+        and data.test_data.dataset.get("masks") is not None
+        and tf.size(data.training_data.dataset["masks"]) > 0  
+        and tf.size(data.test_data.dataset["masks"]) > 0 
+    )
+    else None
+    )
+
+    pix_sampling = model_params.pix_sampling * 1e-6  # Change units from [um] to [m]
+
+    # Ensure star_masks is properly handled
+    star_masks = star_masks if star_masks is not None else None
+
+    reference_shifts = [float(Fraction(value)) for value in model_params.reference_shifts]
+
+    n_stars = len(star_postage_stamps)
+    zernike_centroid_array = []
+
+    # Batch process the stars
+    for i in range(0, n_stars, batch_size):
+        batch_postage_stamps = star_postage_stamps[i:i + batch_size]
+        batch_masks = star_masks[i:i + batch_size] if star_masks is not None else None
+
+        # Compute Zernike 1 and Zernike 2 for the batch
+        zk1_2_batch = -1.0 * compute_zernike_tip_tilt(
+            batch_postage_stamps, batch_masks, pix_sampling, reference_shifts
+        )
+
+        # Zero pad array for each batch and append
+        zernike_centroid_array.append(np.pad(zk1_2_batch, pad_width=[(0, 0), (1, 0)], mode="constant", constant_values=0))
+
+    # Combine all batches into a single array
+    return np.concatenate(zernike_centroid_array, axis=0)
 
 def compute_zernike_tip_tilt(
     star_images: np.ndarray,
