+++ src/wf_psf/tests/test_psf_models/conftest.py
@@ -12,7 +12,7 @@ import pytest
 from wf_psf.utils.read_config import RecursiveNamespace
 from wf_psf.training.train import TrainingParamsHandler
 from wf_psf.psf_models import psf_models
-from wf_psf.data.training_preprocessing import DataHandler
+from wf_psf.data.data_handler import DataHandler
 
 training_config = RecursiveNamespace(
     id_name="_sample_w_bis1_2k",
